Finish GetAvlTime (beta 1)

+++ config/config.go
@@ -7,5 +7,5 @@ type Config struct {
 	ALLOW_ORIGIN  []string `toml:"allow_origin"`
 	SQLITE_FILE   string   `toml:"sqlite_path"`
 	JWT_KEY       string   `toml:"jwt_key"`
-	MAX_QUERY_DAY int64    `toml:"max_query_day"`
+	MAX_QUERY_DAY int      `toml:"max_query_day"`
 }
